requirements.txt
python-telegram-bot==20.7
google-generativeai==0.3.2
python-dotenv==1.0.0
cryptography==41.0.7
imap-tools==1.0.0
python-telegram-bot[job-queue]==20.7